linux-wayland: rename implementation directory
