--- src/platform/linux/audio/capture_stream.rs
@@ -1,8 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct LinuxAudioCaptureConfig {}
-
-impl LinuxAudioCaptureConfig {
-    pub fn new() -> Self {
-        Self {}
-    }
-}
