--- src/platform/linux/audio/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod capture_stream;
-pub(crate) mod frame;
