--- src/platform/linux/mod.rs
@@ -1,18 +1,0 @@
-pub(crate) mod audio;
-pub(crate) mod wayland;
-
-pub(crate) use wayland::capture_stream::WaylandCaptureAccessToken as ImplCaptureAccessToken;
-pub(crate) use wayland::capture_stream::WaylandCaptureConfig as ImplCaptureConfig;
-pub(crate) use wayland::capture_stream::WaylandCaptureStream as ImplCaptureStream;
-pub(crate) use wayland::capture_stream::WaylandPixelFormat as ImplPixelFormat;
-
-pub(crate) use wayland::frame::WaylandVideoFrame as ImplVideoFrame;
-
-pub(crate) use wayland::capture_content::WaylandCapturableApplication as ImplCapturableApplication;
-pub(crate) use wayland::capture_content::WaylandCapturableContent as ImplCapturableContent;
-pub(crate) use wayland::capture_content::WaylandCapturableContentFilter as ImplCapturableContentFilter;
-pub(crate) use wayland::capture_content::WaylandCapturableDisplay as ImplCapturableDisplay;
-pub(crate) use wayland::capture_content::WaylandCapturableWindow as ImplCapturableWindow;
-
-pub(crate) use audio::capture_stream::LinuxAudioCaptureConfig as ImplAudioCaptureConfig;
-pub(crate) use audio::frame::LinuxAudioFrame as ImplAudioFrame;
