--- src/platform/linux/wayland/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod capture_content;
-pub(crate) mod capture_stream;
-pub(crate) mod frame;
