+++ src/platform/linux_wayland/capture_content.rs
