+++ src/platform/linux_wayland/capture_stream.rs
